requirements.txt
asgiref==3.7.2
Django==4.2.6
djangorestframework==3.14.0
GDAL==3.7.2
mysqlclient==2.2.0
phonenumbers==8.13.23
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
